added code injection to help (#876)
