+++ slsDetectorSoftware/generator/commands.yaml
@@ -3145,7 +3145,7 @@ extsig:
       output: [ 'args[0]', "' '", 'args[1]' ]
 
 dac:
-  help: "" # TODO
+  help: "code: return GetHelpDacWrapper(cmd, args);\n" # TODO
   actions:
     GET:
       exceptions:
