+++ slsDetectorSoftware/src/Caller.h
@@ -1,6 +1,7 @@
 // This file is used as input to generate the caller class
 
 #include "CmdParser.h"
+#include "HelpDacs.h"
 #include "sls/Detector.h"
 
 #include <iostream>
